fuse: implement passthrough for getattr/statx
Call vfs_getattr() on backing inode to respond to a user statx(2)
request and update the fuse inode attributes with the response.

For now, we assume that calling vfs_getattr() on backing inode is cheap,
so we never use cached attributed and always update fuse inode attributes
from backing attributes in fuse_permission() and fuse_update_attributes().

Signed-off-by: Amir Goldstein <amir73il@gmail.com>