fuse: support setting backing inode passthrough on getattr
Add backing_id field to getattr out args to allow requesting to
associate a backing inode for inode passthrough operations.

The minimal requirement for the backing inode is to support the
GETATTR passthrough op.

Unlike setting of passthrough from open, this mapping is created for
the entire lifetime of the inode and cannot be changed later, not in
getattr() and not in open() operations.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>